bumped op-rs, moved validated structs to controller/mod.rs, renamed controller.rs -> opa_controller.rs
