--- crate-hashes.json
@@ -1,14 +1,0 @@
-{
-  "git+https://github.com/stackabletech//operator-rs.git?branch=smooth-operator#k8s-version@0.1.3": "0mcr93q8bc18zc2pb9wmri81077qmzw040nivrq656v92am6wi7a",
-  "git+https://github.com/stackabletech//operator-rs.git?branch=smooth-operator#stackable-certs@0.4.0": "0mcr93q8bc18zc2pb9wmri81077qmzw040nivrq656v92am6wi7a",
-  "git+https://github.com/stackabletech//operator-rs.git?branch=smooth-operator#stackable-operator-derive@0.3.1": "0mcr93q8bc18zc2pb9wmri81077qmzw040nivrq656v92am6wi7a",
-  "git+https://github.com/stackabletech//operator-rs.git?branch=smooth-operator#stackable-operator@0.111.1": "0mcr93q8bc18zc2pb9wmri81077qmzw040nivrq656v92am6wi7a",
-  "git+https://github.com/stackabletech//operator-rs.git?branch=smooth-operator#stackable-shared@0.1.1": "0mcr93q8bc18zc2pb9wmri81077qmzw040nivrq656v92am6wi7a",
-  "git+https://github.com/stackabletech//operator-rs.git?branch=smooth-operator#stackable-telemetry@0.6.4": "0mcr93q8bc18zc2pb9wmri81077qmzw040nivrq656v92am6wi7a",
-  "git+https://github.com/stackabletech//operator-rs.git?branch=smooth-operator#stackable-versioned-macros@0.10.0": "0mcr93q8bc18zc2pb9wmri81077qmzw040nivrq656v92am6wi7a",
-  "git+https://github.com/stackabletech//operator-rs.git?branch=smooth-operator#stackable-versioned@0.10.0": "0mcr93q8bc18zc2pb9wmri81077qmzw040nivrq656v92am6wi7a",
-  "git+https://github.com/stackabletech//operator-rs.git?branch=smooth-operator#stackable-webhook@0.9.1": "0mcr93q8bc18zc2pb9wmri81077qmzw040nivrq656v92am6wi7a",
-  "git+https://github.com/stackabletech/krb5-rs.git?tag=v0.1.0#krb5-sys@0.1.0": "148zr0q04163hpirkrff5q7cbxqgwzzxh0091zr4g23x7l64jh39",
-  "git+https://github.com/stackabletech/krb5-rs.git?tag=v0.1.0#krb5@0.1.0": "148zr0q04163hpirkrff5q7cbxqgwzzxh0091zr4g23x7l64jh39",
-  "git+https://github.com/stackabletech/product-config.git?tag=0.8.0#product-config@0.8.0": "1dz70kapm2wdqcr7ndyjji0lhsl98bsq95gnb2lw487wf6yr7987"
-}
